v2.0 Crud entité offer

+++ src/topic/topic.service.ts
@@ -1,6 +1,9 @@
-import { Body, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { GetUser } from 'src/auth/get-user.decorator';
 import { User } from 'src/user/entities/user.entity';
 import { Repository } from 'typeorm';
 import { CreateTopicDto } from './dto/create-topic.dto';
@@ -16,7 +19,7 @@ export class TopicService {
 
   // -----------------------------------------------Méthode afficher tout les TOPICS------------------------//
 
-  async findAll(): Promise<Topic[]> {
+  async findAll() {
     const topicFound = await this.topicRepository.find();
     if (!topicFound) {
       throw new NotFoundException(`Topics non trouvée`);
@@ -25,13 +28,13 @@ export class TopicService {
   }
   // -----------------------------------------------Méthode afficher un TOPIC-------------------------------//
 
-  async findOne(idValue: string): Promise<Topic> {
+  async findOne(idValue: string) {
     try {
-      const topicsFound = await this.topicRepository.findOneBy({
+      const topicFound = await this.topicRepository.findOneBy({
         id: idValue,
       });
       console.log('id du topic----------------', idValue);
-      return topicsFound;
+      return topicFound;
     } catch (error) {
       if (error) {
         throw new NotFoundException(`pas de sujet trouvé avec l'id:${idValue}`);
@@ -40,18 +43,15 @@ export class TopicService {
   }
   // -----------------------------------------------Méthode de création TOPICS-------------------------------//
 
-  async create(
-    @Body() createTopicDto: CreateTopicDto,
-    @GetUser() user: User,
-  ): Promise<Topic> {
+  async create(createTopicDto: CreateTopicDto, user: User) {
     const { title, body, url } = createTopicDto;
     const newTopic = await this.topicRepository.create({
       title,
       body,
       url,
-      userId: user,
+      user,
     });
-
+    console.log('création newTopic-------- ', newTopic);
     try {
       return await this.topicRepository.save(newTopic);
     } catch (error) {
@@ -64,25 +64,33 @@ export class TopicService {
   async update(idValue: string, updateTopicDto: UpdateTopicDto, user: User) {
     //-------------------------Recherche topics dans la BDD -------------------//
 
-    const updateTopicsFound = await this.topicRepository.findOneBy({
+    const topicFound = await this.topicRepository.findOneBy({
       id: idValue,
-      userId: user,
+      user,
     });
     console.log('id requête update topics', idValue);
     console.log('id utilisateur upadate topics', user.id);
 
     //-------------------------Gestion erreur si pas de topic dans la BDD -------//
 
-    if (!updateTopicsFound) {
+    if (!topicFound) {
       throw new NotFoundException("Ce sujet n'existe pas");
+    }
+
+    //-------------------------Gestion erreur si  user pas autorisé -----------//
+
+    if (topicFound.id !== user.id && user.role !== 'admin') {
+      throw new UnauthorizedException(
+        "Vous n'êtes pas autorisé à modifier ces informations",
+      );
     }
 
     //-------------------------Gestion erreur si même valeur-----------//
 
-    if (updateTopicsFound.title === updateTopicDto.title) {
-      throw new Error('Erreur, le titre est le même que preceddment');
+    if (topicFound.title === updateTopicDto.title) {
+      throw new Error('Erreur, le titre est le même que precedemment');
     }
-    if (updateTopicsFound.body === updateTopicDto.body) {
+    if (topicFound.body === updateTopicDto.body) {
       throw new Error('Erreur, le commentaire est le même que precedemment');
     }
     //-----Destructuration de l'update afin de vérifier si données dejà existantes ----//
@@ -91,13 +99,13 @@ export class TopicService {
     console.log('le titre du nouveau commentaire', body);
 
     if (title) {
-      updateTopicsFound.title = title;
+      topicFound.title = title;
     }
     if (body) {
-      updateTopicsFound.body = body;
+      topicFound.body = body;
     }
     try {
-      return await this.topicRepository.save(updateTopicsFound);
+      return await this.topicRepository.save(topicFound);
     } catch (error) {
       throw new Error(`${error}, les données ne sont pas enregistrés`);
     }
@@ -105,15 +113,35 @@ export class TopicService {
 
   // -----------------------------------------------Méthode delete TOPICS by Admin---------------------//
 
-  async remove(id: string, user: User): Promise<Topic | string> {
-    const result = await this.topicRepository.delete({
+  async remove(id: string, user: User) {
+    const topicFound = await this.topicRepository.findOneBy({
       id,
-      userId: user,
     });
-    console.log('résultat du delete par id', result);
-    if (result.affected === 0) {
-      throw new NotFoundException(`pas d'utilisateur trouvé avec l'id:${id}`);
+    console.log('id requête user pour topic', id);
+    console.log('topic trouvé', topicFound);
+
+    //-------------------------Gestion erreur si pas de topic dans la BDD -------//
+
+    if (!topicFound) {
+      throw new NotFoundException("Ce topic n'existe pas");
     }
-    return `Cette action a supprimé le topic de l'utilisateur ${user.nickname}`;
+    //-------------------------Gestion erreur si  user pas autorisé -----------//
+
+    if (topicFound.id !== user.id && user.role !== 'admin') {
+      throw new UnauthorizedException(
+        "Vous n'êtes pas autorisé à modifier ces informations",
+      );
+    }
+    try {
+      const result = await this.topicRepository.delete({
+        id,
+      });
+      console.log('valeur du result par id', result);
+      if (result.affected !== 0) {
+        return `Cette action a supprimé dont le titre était ${topicFound.title}`;
+      }
+    } catch (error) {
+      throw new error(`Impossible de supprimer le user ${error}`);
+    }
   }
 }
